Move ComplexMixIn to plot and subclass in plot3d
